moved the compile code from the DynamicClassLoader to the Writer.
Also renamed MapCodeWriter to MemoryCodeWriter
