+++ src/main/java/com/helger/jcodemodel/inmemory/DynamicClassLoader.java
@@ -135,103 +135,4 @@ public class DynamicClassLoader extends ClassLoader {
     }
   }
 
-  ////
-  // static methods to produce the code in a dynamicclassloader
-  ////
-
-  /**
-   * creates a dynamic class loaders that delegates unknown resources and
-   * classes to the classloader of the javacompiler.
-   */
-  protected static DynamicClassLoader dynCL() {
-    return new DynamicClassLoader(JavaCompiler.class.getClassLoader());
-  }
-
-  /**
-   * generate bytecode in a class loader from a {@link JCodeModel}. This allows
-   * to use, with reflection, the classes defined in the codemodel by loading
-   * them from the class loader.
-   *
-   * @param <T>
-   *          the subtype of DynamicClassLoader
-   * @param cm
-   *          the codemodel to load
-   * @param cl
-   *          the dynamic class loader to add the definitions into
-   * @return cl
-   */
-  public static <T extends DynamicClassLoader> T generate(JCodeModel cm, T cl) {
-    MapCodeWriter codeWriter = new MapCodeWriter();
-    try {
-      new JCMWriter(cm).build(codeWriter);
-    } catch (IOException e) {
-      throw new UnsupportedOperationException("catch this exception", e);
-    }
-    return generate(codeWriter.getBinaries().entrySet(), cl);
-  }
-
-  /**
-   * shortcut for {@link #generate(JCodeModel, DynamicClassLoader)} with a
-   * correct classloader
-   */
-  public static DynamicClassLoader generate(JCodeModel cm) {
-    return generate(cm, dynCL());
-  }
-
-  /**
-   * generate bytecode in a class loader from an iterable over source files.
-   * This allows to use, with reflection, the classes defined by their sources,
-   * by loading them from the class loader.
-   *
-   * @param <T>
-   *          the subtype of DynamicClassLoader
-   * @param sourceFiles
-   *          the source files to load
-   * @param cl
-   *          the dynamic class loader to add the definitions into
-   * @return cl
-   */
-  public static <T extends DynamicClassLoader> T generate(Iterable<Entry<String, ByteArrayOutputStream>> sourceFiles,
-      T cl) {
-    ArrayList<JavaFileObject> compilationUnits = new ArrayList<>();
-    HashMap<String, ByteArrayOutputStream> nonJava = new HashMap<>();
-    for (Entry<String, ByteArrayOutputStream> e : sourceFiles) {
-      if (e.getKey().endsWith(".java")) {
-        try {
-          compilationUnits.add(new SourceJavaFile(e.getKey(), e.getValue().toString()));
-          String className = e.getKey().replaceAll("/", ".").replace(".java", "");
-          CompiledCodeJavaFile cc = new CompiledCodeJavaFile(className);
-          cl.setCode(cc);
-        } catch (Exception e1) {
-          throw new UnsupportedOperationException("catch this exception", e1);
-        }
-      } else {
-        nonJava.put(e.getKey(), e.getValue());
-      }
-    }
-    if (!compilationUnits.isEmpty()) {
-      try {
-        ForwardingJavaFileManager<JavaFileManager> fileManager = new ClassLoaderFileManager(
-            javac.getStandardFileManager(diagnostic -> System.err.println("file diagnostic " + diagnostic), null, null),
-            cl);
-        JavaCompiler.CompilationTask task = javac.getTask(null, fileManager,
-            diagnostic -> System.err.println(" compile diagnostic " + diagnostic), null, null, compilationUnits);
-        task.call();
-      } catch (Exception e1) {
-        throw new UnsupportedOperationException("catch this exception", e1);
-      }
-    }
-    cl.addResources(nonJava);
-    return cl;
-  }
-
-  /**
-   * shortcut for {@link #generate(Iterable, DynamicClassLoader)} with a correct
-   * class loader
-   */
-  public static DynamicClassLoader generate(Iterable<Entry<String, ByteArrayOutputStream>> sourceFiles) {
-    return generate(sourceFiles, dynCL());
-  }
-
-  private static final JavaCompiler javac = ToolProvider.getSystemJavaCompiler();
 }
